Add files via upload
This includes a copy of FILECFG/95's source and a ZIP file when Robert King updated it for me, John Riley, in 2010.  We renamed it FileCFG/32.  Robert King gave me a copy of the source code, but I've lost it.  I will add more Windows 32Bit ProBoard Utilities here as they become available to me.